Update placehold.html
Added a line to say that if it is an error to contact our support team and changed the tittle to currently under construction. If you want to add me to the made by, you can, I don't really care because I did not make the page, I only edited it. If you want to change it back, go ahead. I'm only suggesting this due to the fact of if somethings broken then the person knows to contact our support team. - Kristopher Mencarelli