Update 404.html
Added a line that says to contact support if they think it was a bug, added 2 words to the choose kindness button. We just have to add the 3 team members who are missing which I'm not going to go crazy with right now due to the issue with the team page. If you want to revert my changes, go ahead, I don't care because I am just placing this as a suggestion. - Kristopher Mencarelli.